Output vrml in double precision

+++ include/dime/Basic.h
@@ -42,8 +42,8 @@
 
 // we prefer to use floats to save mem. Applications needing
 // scientific calculations should typedef this to double
-typedef float dxfdouble;
-// typedef double dxfdouble;
+//typedef float dxfdouble;
+typedef double dxfdouble;
 
 #include <float.h>
